# Performance improvements to configuration model for simple graph